backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
trimesh==4.1.0
numpy==1.26.3
scipy>=1.11.0
networkx>=3.0
openai==1.12.0
shapely>=2.0.0
pyvista>=0.46.0
meshio>=5.3.0